scraper/retailers/retailer_a.py
"""
retailer_a.py — Scraping logic for Retailer A.

Implement the search() function below. It should return a list of dicts
with keys: product_name, brand, package_size, price, unit_price, category.

The retailer name and scrape_date are added by the main scraper.py orchestrator.
"""

import requests
from bs4 import BeautifulSoup
from scraper.utils import parse_price, parse_unit_price

RETAILER_NAME = "retailer_a"
BASE_URL = "https://www.example-retailer-a.com"  # Replace with actual URL

CATEGORY_MAP = {
    "milk":              "dairy",
    "eggs":              "dairy",
    "bread":             "bakery",
    "bananas":           "produce",
    "chicken breast":    "meat",
    "ground beef":       "meat",
    "cereal":            "dry_goods",
    "pasta":             "dry_goods",
    "rice":              "dry_goods",
    "canned vegetables": "canned_goods",
}


def search(product: str, headers: dict) -> list[dict]:
    """
    Search for a product and return a list of result dicts.

    Args:
        product: Product name to search (e.g. "milk")
        headers: HTTP headers from utils.get_headers()

    Returns:
        List of dicts with keys:
            product_name, brand, package_size, price, unit_price, category
    """
    # TODO: implement retailer-specific search logic
    # Stub returns empty list until implemented
    results = []

    try:
        # Example structure — adapt to actual retailer HTML:
        # url = f"{BASE_URL}/search?q={product.replace(' ', '+')}"
        # resp = requests.get(url, headers=headers, timeout=15)
        # resp.raise_for_status()
        # soup = BeautifulSoup(resp.text, "lxml")
        # for item in soup.select(".product-card"):
        #     name  = item.select_one(".product-name").text.strip()
        #     brand = item.select_one(".product-brand").text.strip()
        #     size  = item.select_one(".product-size").text.strip()
        #     price_raw = item.select_one(".product-price").text.strip()
        #     unit_raw  = item.select_one(".unit-price").text.strip()
        #     results.append({
        #         "product_name": product,
        #         "brand":        brand,
        #         "package_size": size,
        #         "price":        parse_price(price_raw),
        #         "unit_price":   parse_unit_price(unit_raw),
        #         "category":     CATEGORY_MAP.get(product, "other"),
        #     })
        pass

    except Exception as e:
        print(f"  [{RETAILER_NAME}] Error on '{product}': {e}")

    return results
