requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
xgboost>=2.0
requests>=2.31
beautifulsoup4>=4.12
lxml>=4.9
selenium>=4.15
fake-useragent>=1.4
pydantic>=2.0
matplotlib>=3.7
seaborn>=0.12
jupyter>=1.0
joblib>=1.3
pyyaml>=6.0
